Add GLFW and clean up build system
Added the source code to GLFW. It will be compiled WITH our application
rather than me just linking a glfw3.lib or libglfw3.a file. This is just
for a cleaner build system (we dont have to have separate libraries for
both windows and linux now, and the build config of GLFW will now always match
the build config of our application, which is neat.