Updated anti-aliasing & changed particles to be green
Anti Aliasing has been set to only affect the outermost pixel of
circles, since the more you anti alias the smaller your circles
appear to be, and this was causing it to seem as if there was a
gap between the particles. The particles have also been set to be
green as they are being made to reprisent tennis balls.