Update particle collision
When moving the particles away from eachother along their velocities,
the "offset" amount was causing the simulation to become unstable after
20 seconds. This was because there were actually two versions of the
offset, and either one of them was correct depending on the situation.
I added code for finding which the correct offset is. Now the simulation
is stable for a minute and 14 seconds. More bugs to squash!!!