Edit Build System
Added the include directory for GLFW to the imgui CMakeLists.txt file.
ImGui was able to access the GLFW header file on linux but not on
windows because of the differences in compilers. This is now fixed. Also
edited how the appearances window behaves. It is now locked to the corner of
the screen and cannot be dragged or resized