Added Shape, Circle, PhysObject & World classes
Added a base class "Shape", and a class which is derived from the shape class,
"Circle". There will also be a "Polygon" class in the future. Added a "PhysObject"
class which stores a "Shape" class and the position, old_position, mass of our
physics object. There is also a world class which can store many PhysObject
classes.