Fixed Newton's Cradle Scene
When using Newton's Cradle Scene with a high substep count, the springs
would freak out and explode. This was because some of the particles
were being spawned very far away from where they were meant to spawn,
and so the springs would freak out and pull them extremely quickly back
to where they were meant to be.