Backup before insanely restructuring code
I need to restructure how the world take a step forward in time because
Runge-Kutta 4 depends on knowing the acceleration of a particle at any
given moment in time. This is not good when it comes to external forces
on the particles which update based on not only the particle's position,
but also the positions of every other particle in the world which could
be attached to these particles by a spring. Therefore we need to udpate
the entire world at the same time when calculating the k1, k2, k3 and k4
values of each particle so that the acceleration of each particle is
accurate. Why do i have to make my life so complicated hahahha