Fixed static particle velocity bug
When a particle was static, its velocity would remain exactly the same
even if it was not being moved by the cursor. This meant that when a
particle was turned static after moving with a high velocity, this
velocity would remain as the particle's velocity even though it was
static, resulting in wacky unrealistic collisions. Now after every
collision, the static particle's velocity is reset, fixing the issue
completely