No longer reset static particles' velocities after collision
The fix for the bug where particles' velocities would remain
after they had become static broke another part of the program. When
moving a static particle with the mouse, the velocity of the particle
would only affect one other particle in the scene before being reset.
That was not the proper fix to the solution.

I tried resetting the velocity of every static particle after every
collision had been resolved in a step, but this meant that you could no
longer throw particles with the cursor.

This will need some thought put into it. For now the only solution i see
is to implement verlet integration, but that requires a reworking of the
engine's core mechanics and i do not feel up to the task yet.